move workflows into their own tab
