+++ indigo_app/templates/place/tabbed_layout.html
@@ -35,6 +35,9 @@
       <a class="nav-item nav-link {% if view.tab == 'tasks' %}active{% endif %}" href="{% url 'tasks' place=place.place_code %}">
         Tasks <span class="badge badge-light">{{ place.tasks.unclosed.count }}</span>
       </a>
+      <a class="nav-item nav-link {% if view.tab == 'workflows' %}active{% endif %}" href="{% url 'workflows' place=place.place_code %}">
+        Workflows <span class="badge badge-light">{{ place.workflows.count }}</span>
+      </a>
       <a class="nav-item nav-link {% if view.tab == 'activity' %}active{% endif %}" href="{% url 'place_activity' place=place.place_code %}">Activity</a>
     </nav>
     {% endblock %}
