+++ indigo_app/urls.py
@@ -1,7 +1,7 @@
 from django.conf.urls import url, include
 from django.views.generic.base import RedirectView, TemplateView
 
-from .views import users, works, documents, tasks, places
+from .views import users, works, documents, tasks, places, workflows
 
 
 urlpatterns = [
@@ -30,9 +30,10 @@ urlpatterns = [
     url(r'^places/(?P<place>[a-z]{2}(-[^/]+)?)/tasks/(?P<pk>\d+)/reopen', tasks.TaskChangeStateView.as_view(change='reopen'), name='reopen_task'),
     url(r'^places/(?P<place>[a-z]{2}(-[^/]+)?)/tasks/(?P<pk>\d+)/unsubmit', tasks.TaskChangeStateView.as_view(change='unsubmit'), name='unsubmit_task'),
     url(r'^places/(?P<place>[a-z]{2}(-[^/]+)?)/tasks/(?P<pk>\d+)/close', tasks.TaskChangeStateView.as_view(change='close'), name='close_task'),
-    url(r'^places/(?P<place>[a-z]{2}(-[^/]+)?)/workflows/new$', tasks.WorkflowCreateView.as_view(), name='workflow_create'),
-    url(r'^places/(?P<place>[a-z]{2}(-[^/]+)?)/workflows/(?P<pk>\d+)/$', tasks.WorkflowDetailView.as_view(), name='workflow_detail'),
-    url(r'^places/(?P<place>[a-z]{2}(-[^/]+)?)/workflows/(?P<pk>\d+)/edit$', tasks.WorkflowEditView.as_view(), name='workflow_edit'),
+    url(r'^places/(?P<place>[a-z]{2}(-[^/]+)?)/workflows/$', workflows.WorkflowListView.as_view(), name='workflows'),
+    url(r'^places/(?P<place>[a-z]{2}(-[^/]+)?)/workflows/new$', workflows.WorkflowCreateView.as_view(), name='workflow_create'),
+    url(r'^places/(?P<place>[a-z]{2}(-[^/]+)?)/workflows/(?P<pk>\d+)/$', workflows.WorkflowDetailView.as_view(), name='workflow_detail'),
+    url(r'^places/(?P<place>[a-z]{2}(-[^/]+)?)/workflows/(?P<pk>\d+)/edit$', workflows.WorkflowEditView.as_view(), name='workflow_edit'),
     url(r'^places/(?P<place>[a-z]{2}(-[^/]+)?)/works/new/$', works.AddWorkView.as_view(), name='new_work'),
     url(r'^places/(?P<place>[a-z]{2}(-[^/]+)?)/works/new-batch/$', works.BatchAddWorkView.as_view(), name='new_batch_work'),
