+++ indigo_app/views/tasks.py
@@ -236,64 +236,3 @@ class TaskChangeStateView(TaskViewBase, View, SingleObjectMixin):
         task.save()
 
         return redirect('task_detail', place=self.kwargs['place'], pk=self.kwargs['pk'])
-
-
-class WorkflowCreateView(TaskViewBase, CreateView):
-    # permissions
-    permission_required = ('indigo_api.add_workflow',)
-
-    js_view = ''
-
-    context_object_name = 'workflow'
-    form_class = WorkflowForm
-    model = Workflow
-
-    def get_context_data(self, *args, **kwargs):
-        context = super(WorkflowCreateView, self).get_context_data(**kwargs)
-
-        context['place_open_tasks'] = self.place.tasks.filter(state__in=Task.OPEN_STATES)
-
-        return context
-
-    def get_form_kwargs(self):
-        kwargs = super(WorkflowCreateView, self).get_form_kwargs()
-
-        workflow = Workflow()
-        workflow.country = self.country
-        workflow.locality = self.locality
-        workflow.created_by_user = self.request.user
-
-        kwargs['instance'] = workflow
-
-        return kwargs
-
-    def get_success_url(self):
-        return reverse('workflow_detail', kwargs={'place': self.kwargs['place'], 'pk': self.object.pk})
-
-
-class WorkflowDetailView(TaskViewBase, DetailView):
-    context_object_name = 'workflow'
-    model = Workflow
-
-
-class WorkflowEditView(TaskViewBase, UpdateView):
-    # permissions
-    permission_required = ('indigo_api.change_workflow',)
-
-    context_object_name = 'workflow'
-    form_class = WorkflowForm
-    model = Workflow
-
-    def form_valid(self, form):
-        self.object.updated_by_user = self.request.user
-        return super(WorkflowEditView, self).form_valid(form)
-
-    def get_context_data(self, *args, **kwargs):
-        context = super(WorkflowEditView, self).get_context_data(**kwargs)
-
-        context['place_open_tasks'] = self.place.tasks.filter(state__in=Task.OPEN_STATES)
-
-        return context
-
-    def get_success_url(self):
-        return reverse('workflow_detail', kwargs={'place': self.kwargs['place'], 'pk': self.object.pk})
